Prevent fatal errors from leaking
The errors generated during feature detection are sometimes picked up by
frameworks or libraries that use `error_get_last()` in a shutdown
handler to look for fatal errors.

This change forces an `E_NOTICE` error to be triggered when an error may
potentially be leaked, preventing false positives in *most* situations.

Fixes #209